Move tests to main test file
