+++ test.ts
@@ -237,6 +237,26 @@ test('prototype support', t => {
 	t.is(unicorn.foo(), 0);
 });
 
+test('.decorator()', (t: ExecutionContext) => {
+	class TestClass {
+		constructor() {
+			this.i = 0;
+		}
+
+		@mem.decorator()
+		counter() {
+			return ++this.i;
+		}
+	}
+
+	const alpha = new TestClass();
+	t.is(alpha.counter(), 1);
+	t.is(alpha.counter(), 1, 'The method should be memoized');
+
+	const beta = new TestClass();
+	t.is(beta.counter(), 1, 'The method should not be memoized across instances');
+});
+
 test('mem.clear() throws when called with a plain function', t => {
 	t.throws(() => {
 		mem.clear(() => {});
